Fix comment bug
