+++ src/styles/ChallengeCard.css
@@ -1,11 +1,12 @@
 .show-single-card {
   border: none;
-  border-radius: 100%;
+  /*border-radius: 100%;*/
   color: gray;
   text-decoration: none;
-  font-size: 20px;
+  font-size: 25px;
   float: right;
-  margin: 12px;
+  margin: 8px;
+  background: none;
 }
 
 .show-single-card:hover{
@@ -45,12 +46,12 @@
 .edit-btn {
   position: absolute;
   bottom: 0;
-  right: 0;
+  left: 0;
   font-size: 30px;
   border: none;
   color: gray;
   background: none;
-  margin: 0 40px 5px 0;
+  margin: 0 40px 5px 8px;
 }
 
 .edit-btn:hover {
@@ -59,6 +60,8 @@
 }
 
 .checked {
+  width: 40px;
+  height: 40px;
   position: absolute;
   font-size: 25px;
   border: 1px solid gray;
