+++ src/styles/ChallengesList.css
@@ -23,8 +23,8 @@
 }
 
 .challenge-elements:hover {
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.9);
-  border-radius: 5px;
+  /*box-shadow: 0 0 10px rgba(0, 0, 0, 0.9);
+  border-radius: 5px;*/
 }
 
 .single-challenge-image {
